+++ relayer/src/event.rs
@@ -1,2 +1,3 @@
 pub mod bus;
 pub mod monitor;
+pub mod rpc;
